+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Denvig Changelog
 
 
-## [Unreleased]
+## [0.3.0] - 2025-09-24
 
 ### Added